+++ bugsnag-android-core/src/main/java/com/bugsnag/android/internal/ImmutableConfig.kt
@@ -220,7 +220,7 @@ internal fun sanitiseConfiguration(
 
     @Suppress("SENSELESS_COMPARISON")
     if (configuration.delivery == null) {
-        configuration.delivery = DefaultDelivery(connectivity, configuration.logger!!)
+        configuration.delivery = DefaultDelivery(connectivity, configuration.apiKey, configuration.logger!!)
     }
     return convertToImmutableConfig(
         configuration,