+++ bugsnag-android-core/src/main/java/com/bugsnag/android/internal/InternalMetrics.kt
@@ -16,4 +16,11 @@ interface InternalMetrics {
     fun notifyAddCallback(callback: String)
 
     fun notifyRemoveCallback(callback: String)
+
+    fun setBreadcrumbTrimMetrics(breadcrumbsRemoved: Int, bytesRemoved: Int)
 }
+
+internal data class TrimMetrics(
+    val itemsTrimmed: Int, // breadcrumbs, strings, whatever
+    val dataTrimmed: Int // chars, bytes, whatever
+)