+++ bugsnag-android-core/src/test/java/com/bugsnag/android/DeliveryTest.kt
@@ -7,7 +7,7 @@ class DeliveryTest {
 
     @Test
     fun testResponseCodeMapping() {
-        val delivery = DefaultDelivery(null, NoopLogger)
+        val delivery = DefaultDelivery(null, "myApiKey", NoopLogger)
         assertEquals(DeliveryStatus.DELIVERED, delivery.getDeliveryStatus(202))
         assertEquals(DeliveryStatus.UNDELIVERED, delivery.getDeliveryStatus(503))
         assertEquals(DeliveryStatus.UNDELIVERED, delivery.getDeliveryStatus(0))